Add vue config
This patch adds a starting point of a working vue with typescript
configuration. It externalizes vue, vue-class and vue-property-decorator.
Also it adds the necessary tsconfig.
It also extends the docker-compose.yml in terms of a alignment with
the current bridge version.
It also separates the build steps into 2 parts and adds for calling them
both npm-run-all.

Change-Id: I50e70c4cadef83dc3c666f3c601ebaf0e77c054f