Add radio input component
Vertical alignment and grouping is left to the consuming application.

The styles from this component are largely adapted from:
https://github.com/wikimedia/oojs-ui/blob/a0cc9490a361af3fe18f91f16ae7abe5672e8bfd/src/themes/wikimediaui/widgets.less#L1130
They were somewhat reordered to make use of nesting in SCSS.
Also, workarounds for old browsers were removed, because we are using
auto-prefixing here.

The label receives `white-space: normal` to make it wrap again.

It also receives a padding-top: 0.1em; to align the label text properly
with the circle. This is somewhat of a smelly hack, but the alternative
would be to use align-items: center on the top div, position the
description absolutely and somehow magically increase the height of top
div to still include the description.

Some aspects cannot be tested with unit-tests, due to the limitations of
jest, jsdom and @vue/test-utils.
We might need a different setup to assert that it
- does not emit on deselect
- selects on a click on the label
- selects on a click on the description
- doesn't do anything when disabled

The slots in the info tab of storybook lack their name. It is unclear
why that is. 🤷

Bug: T239799
Change-Id: Idb89411c53f70ec06a1e6bff52508a956c7401dc