Fix font sizes of RadioInput
In the Data Bridge, we want the label to be 16px large and the
description at 14px, rather than 14px and 12px. More generally, though,
the surrounding component should be able to control this, so change the
label font size to 1em (the surrounding component can then set the whole
RadioInput to 1rem for 16px, 0.875rem for 14px, etc.) and make the
description relative to that.

The component doesn’t yet work very well for font sizes that are very
different from 16px (the vertical alignment of the radio button and the
label text gets broken), but that is left for later.

Bug: T244991
Change-Id: Ibde7afd8f84b437ce5e0dd17a2381907f1ec250d