Make RadioInput work at any font size
Center the input circle (more or less) relative to the main label,
regardless of the main label’s font size (which can be controlled by the
surrounding page). The rationale is as follows:

* The center of the main label’s text is more or less the center of its
  content box.
* The main label’s content box has the height of its font-size times its
  line-height.
* The main label’s content box has the same top edge as the surrounding
  div for the whole radio input, which is the base top edge of the input
  circle as well. (But see below.)
* Thus, for the center of the input circle to be at the same height as
  the center of the main label’s text, it must be offset from the top by
  half the height (font-size times line-height) of the main label minus
  half the height of the input circle.

Because this calculation mixes em (main label) and rem (input circle),
it must be done in the browser, using calc(). Fortunately, that is
supported since IE9 according to MDN [1].

The calculation (a string as far as SCSS is concerned) is extracted into
a variable so that it can be applied both to the invisible <input> and
the <span> that actually renders the circle. The <input>’s font-size
also needs to be set to 1em so it does the calculation correctly.

The line-height of the main label (and, for consistency, of the
description as well) is extracted into a variable so it can be used in
the calculation. It is also specified as a unitless decimal, rather than
a percentage – they seem to produce the same visual result, but SCSS can
only calculate the font-size * line-height product if the line-height is
a decimal, not a percentage. MDN also recommends using numbers, not
percentages or lengths, for line-height [2].

Bullet point three of the above rationale notwithstanding, the “minor
adjustment” (code) or “smelly hack” (commit message of 9dfe41cc06) of
the padding-top: .1em on the whole label (main label and description) is
retained. The effect is that the input circle is in fact not *quite*
centered relative to the main label’s content box (which, along with the
description, is offset by those .1em); however, at all three font sizes
now shown in the storybook, this appears to us to look nicer than the
“more correct” version without the padding-top.

[1]: https://developer.mozilla.org/en-US/docs/Web/CSS/calc#Browser_compatibility
[2]: https://developer.mozilla.org/en-US/docs/Web/CSS/line-height

Bug: T244991
Change-Id: Ia2ec06740279a59e19aa1a220d41dfa8ca199aca