Add (untested) GitHub workflow to publish on tag
This will be used to automatically publish the package to npmjs.org when
a new tag is pushed (usually mirrored from Gerrit), provided that the
NPM_AUTH_TOKEN secret is configured. However, to test the workflow, we
first run the publish in dry-run mode, and trigger the workflow on every
push, not just tags. That will be cleaned up afterwards.

Based on the config from wmde/eslint-config-wikimedia-typescript [1],
with the following changes (besides the test things mentioned above):

- Trigger on tag push, rather than created release, as this repository
  is primarily hosted on Gerrit and GitHub is merely a mirror for it.
- Name the checkout step, so all steps are named.
- Include version of checkout action, as recommended [2].
- Remove the NODE_AUTH_TOKEN environment variable, as it does not appear
  to be used.
- Use printf rather than echo to write the ~/.npmrc file, to prevent
  echo interpreting any escape sequences.
- Single-quote the auth token, as we don’t want the shell to expand it.

[1]: https://github.com/wmde/eslint-config-wikimedia-typescript/blob/b240d5d09c7bfc02fc87fae98353eb15a2b63c84/.github/workflows/release.yml
[2]: https://help.github.com/en/actions/reference/workflow-syntax-for-github-actions#jobsjob_idstepsuses

Bug: T245497
Change-Id: I2179b0f5f21b7e8ac298951edb410e400d713075