Resize ResizingTextField on window resize
When the window is resized, the needed height of the text area might
change, so we should recalculate it. (It may also make sense to
recalculate it when the element itself resizes for some reason, but
apparently there’s no stable API for this yet.)

A story to simplify testing this is included.

Bug: T246765
Change-Id: Ie0d64f97add4d859c7904d2b75e17a26a234dd48