Make RadioInput work in IE11
IE11 doesn’t fire an “input” event on radio inputs. It does fire “click”
and “change”, and both of these are also fired on other browsers
(including when there is no real mouse click, i. e. if the input was
selected using the keyboard: the browser fakes a click event). “click”
also fires when the already selected input is clicked again, whereas
“change” only fires on, well, a change, so use that event.

Note that we still emit an “input” event on the Vue side, so this is not
a breaking change for users of the component.

(Also fixes a mistake in a test name – I assume “input element” was
meant to be “input event” all along, now it’s “change event”.)

Bug: T249343
Change-Id: I529bd5edad6acd46d6dd1561f611ab0ea2798ed2