storybook: smoke test it
Validate the storybook config as part of our test run - to ensure there
is a fair chance of having a functioning storybook. (Does not prevent
malfunctioning stories however).

This was first added to bridge in 0dde935 (wikibase).

This project is built on blubber which uses a runuser with restricted
privileges - it can not write to node_modules which conflicts with
storybook's attempts to store cache there. CACHE_DIR overrides the
directory to store the cache inside of.

I filed https://phabricator.wikimedia.org/T255434 to get clarity about
the "right" approach with the blubber maintainers.

Change-Id: If5258202bf14a781dfef922316c91efee6815e6a